refactor toogle logic
